requirements.txt

torch
torchvision
streamlit
pillow
numpy
matplotlib
scikit-learn
